Restore run_command after processes have drained
When the burst buffer plugin is unloaded run_command drains all
running processes. However, it never resets run_command to work
as normal again.

This casues issues when you use Slurm in HA mode, the backup
server only works for the first failover, but if it gets used
again, all burst buffer commands get killed, and the output of
the command is always ignored. This shows its self in the logs
as a failed parse of the pools output.

Bug 8233